Rename Ranker to Game
I decided that after re-reading the guidelines a million times it would be most appropriate to just have a Game class and I think instead of thinking of hands as their own object, the players will just be represented by the hands. I'm really unsure if there's a situation where a player won't have a hand, but I don't think there is for now.

I might have to play some online poker to learn how this works.